Update library to 2.4.0. iOS 8, Xcode 6, 1Password.

+++ SampleApp/PayPal-iOS-SDK-Sample-App/ZZFlipsideViewController.m
@@ -38,7 +38,7 @@
 - (void)viewWillAppear:(BOOL)animated {
   [self logEnvironment];
 
-  int numberOfSegments = self.environmentSegmentedControl.numberOfSegments;
+  int numberOfSegments = (int)self.environmentSegmentedControl.numberOfSegments;
   while (numberOfSegments--) {
     NSString *title = [self.environmentSegmentedControl titleForSegmentAtIndex:numberOfSegments];
     if ([[title lowercaseString] isEqualToString:self.delegate.environment]) {
